% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markets18}
\alias{markets18}
\title{Daily trading prices of 112 competitive 2018 Midterm races}
\format{
A data frame with 41,933 rows and 9 variables:
\describe{
\item{date}{Prediction date}
\item{mid}{Market ID}
\item{race}{Unique race code}
\item{party}{Contract party}
\item{open}{Opening contract price for day}
\item{low}{Low contract price for day}
\item{high}{High contract price for day}
\item{close}{Closing contract price for day}
\item{volume}{Nunber of contract shares traded}
...
}
}
\source{
\url{https://www.predictit.org/}
}
\usage{
markets18
}
\description{
The opening, low, high, and closing prices of contracts tied to a party
winning a midterm race. Equilibrium prices reflect outcome probability.
}
\details{
Prediction markets generate probabilistic forecasts by crowd-sourcing the
collection of data from self-interested and risk averse traders. The
efficient market hypothesis holds that asset prices reflect all available
information (including forecasting models).

PredictIt is an exchange run by Victoria University of Wellington, New
Zealand. The site offers a continuous double-auction exchange, where traders
buy and sell shares of futures contracts tied to election outcomes. As a
trader’s perception of probabilities changes, they can sell those shares. The
market equilibrium price updates accordingly to reflect current probability.
As outcomes become more likely, prices rise as demand for shares increases.
Market data for all midterm prices was provided for academic purposes.
}
\keyword{datasets}
