% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{members115}
\alias{members115}
\title{Members of the 115th Congress}
\format{
A data frame with 534 rows and 9 variables:
\describe{
\item{name}{Legislator full name}
\item{gid}{Unique GovTrack ID number}
\item{birth}{Birth date}
\item{race}{District or senate state and class code}
\item{party}{Political party}
\item{gender}{Identifying gender}
\item{chamber}{Chamber of Congress: House or Senate}
\item{ideology}{Left-Right (0-1) ideology score from pattern of cosponsorship}
\item{leadership}{Leadership score based on frequency of cosponsored bills}
...
}
}
\source{
\url{https://theunitedstates.io/congress-legislators}
}
\usage{
members115
}
\description{
The biographical information of the members is taken from an archived
version of the \code{current-legislators} data from the UnitedStates.io project.
The ideology and leadership scores are taken from GovTrack, which bases their
analysis on the bills introduced and cosponsored by members.
}
\keyword{datasets}
