% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model18}
\alias{model18}
\title{Daily forecast probabilities of all 2018 Midterm races}
\format{
A data frame with 1,049 rows and 11 variables:
\describe{
\item{date}{Prediction date}
\item{race}{Unique race code}
\item{name}{Candidate name}
\item{party}{Candidate party}
\item{chamber}{Chamber of Congress: House or Senate}
\item{special}{Is this election a special election?}
\item{incumbent}{Is the candidate the incumbent?}
\item{prob}{Probability candidate wins}
\item{voteshare}{Median estimated share of vote}
\item{min_share}{10th percentile vote share}
\item{max_share}{90th percentile vote share}
...
}
}
\source{
\url{https://fivethirtyeight.com/}
}
\usage{
model18
}
\description{
The election forecast probabilities from the 2018 FiveThirtyEight "classic"
midterm model. Probabilities are taken by estimating vote share and a
probability distribution from historical data.
}
\details{
As Nate Silver explains, most forecasting models (1) “take lots of polls,
perform various types of adjustments to them, and then blend them with other
kinds of empirically useful indicators to forecast each race”. Importantly,
they (2) “account for the uncertainty in the forecast and simulate the
election thousands of times” to generate a probabilistic forecast.

The classic model incorporates three types of inputs, primarily direct and
imputed polling as well as fundamentals factors like incumbency and the
generic ballot.
}
\keyword{datasets}
