% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memento.R
\name{read_memento}
\alias{read_memento}
\title{Read a file from the Time Travel API}
\source{
\url{https://github.com/hrbrmstr/wayback/blob/master/R/mementoweb.r}
}
\usage{
read_memento(
  url,
  date = format(Sys.Date(), "\%Y"),
  as = c("text", "raw", "parsed"),
  ...
)
}
\arguments{
\item{url}{URL to retrieve information for}

\item{date}{Date to use when checking for availability. If you don't
pass in a valid R "time-y" object, you will need to ensure the character
string you provide is in a valid subset of \code{YYYYMMDDhhmmss}.}

\item{as}{How you want the content returned. One of "\code{text}", "\code{raw}" or
"\code{parsed}" (it uses \code{\link[httr:content]{httr::content()}} to do the heavy lifting).}

\item{...}{Additional arguments passed to \code{\link[httr:content]{httr::content()}}.}
}
\value{
The content from the GET request.
}
\description{
This function is adapted from the hrbrmstr/wayback package on GitHub.
}
