% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results18}
\alias{results18}
\title{2018 Midterm race results}
\format{
A data frame with 468 rows and 6 variables:
\describe{
\item{year}{Election year}
\item{race}{Unique race code}
\item{lean}{The average difference between how a state or district votes
and how the country votes overall}
\item{category}{Ordinal race category: safe > solid > likely > lean > tilt}
\item{winner}{\code{TRUE} if Democrat won, \code{FALSE} if Republican}
\item{party}{Incumbent party}
...
}
}
\source{
\url{https://fivethirtyeight.com/}
}
\usage{
results18
}
\description{
2018 Midterm race results
}
\keyword{datasets}
