% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-memento.R
\name{write_memento}
\alias{write_memento}
\title{Write a memento file}
\usage{
write_memento(url, date, dir)
}
\arguments{
\item{url}{URL to retrieve information for.}

\item{date}{The date of the archive.}

\item{dir}{The directory in which to save the file.}
}
\value{
The path of the created file, invisibly.
}
\description{
Read an archived file and write it to a local directory.
}
